+++ src/main/java/com/tansoflow/tansocore/repository/ExternalApiKeyRepository.java
@@ -22,10 +22,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.UUID;
 
 @Repository
 public interface ExternalApiKeyRepository extends JpaRepository<ExternalApiKey, UUID>  {
     @Query(value = "SELECT eak FROM ExternalApiKey eak WHERE eak.keyType = :keyType AND eak.account = :account")
     ExternalApiKey findExternalApiKeyByKeyTypeAndAccount(String keyType, UUID account);
+
+    List<ExternalApiKey> findAllByKeyType(String keyType);
 }